big patch 16-12-15

+++ database/migrations/2015_11_27_144954_create_deal_table.php
@@ -14,7 +14,7 @@ class CreateDealTable extends Migration
     {
         Schema::create('deals', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('qty');
+            $table->decimal('qty', 6, 2);
             $table->decimal('amount', 10, 2);
             $table->timestamps();
